+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Backport static code diagnostic `member-ordering-extended`.
+- Backport static code diagnostics `member-ordering-extended`, `avoid-returning-widgets`.
 - Backport fix excludes on Windows OS.
 
 ## 2.4.1